// packages/commons-service/src/types/ProductShoe.ts
export interface ProductShoe {
  id: string;
  dwid?: string;
  brand: string;
  year?: number;
  month?: number;
  day?: number;
  title: string;
  subtitle?: string;
  url: string;
  image?: string;
  priceSale: number;
  priceOriginal: number;
  gender?: string;
  ageGroup?: string;
  sizes?: string[] | null;
}